ddp_backend/pyproject.toml
[project]
name = "ddp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.6",
    "insightface>=0.7.3",
    "matplotlib>=3.10.8",
    "nest-asyncio>=1.6.0",
    "numpy>=2.4.2",
    "onnxconverter-common>=1.16.0",
    "onnxruntime-gpu>=1.24.1; sys_platform == 'linux'",
    "onnxruntime>=1.24.1; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "onnxruntime>=1.18.0,<1.21.0; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "pyngrok>=7.5.0",
    "pywavelets>=1.9.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
    "stt",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "types-pyyaml>=6.0.12.20250915",
    "unite-detection[gpu-no-flash-attn]",
    "uvicorn>=0.40.0",
    "wavelet-clip",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "apscheduler>=3.11.2",
    "taskiq>=0.12.1",
    "taskiq-redis>=1.2.2",
    "taskiq-fastapi>=0.4.0",
    "pydantic>=2.12.5",
    "redis>=7.2.0",
    "types-redis>=4.6.0.20241004",
    "sqlmodel>=0.0.37",
]
